chore: move components to lib folder
