+++ react/lib/assets/BCH.svg.txt
