+++ react/lib/assets/edit-pencil.tsx
