+++ react/lib/components/BarChart/BarChart.stories.tsx
