+++ react/lib/components/BarChart/BarChart.tsx
