+++ react/lib/components/BarChart/index.ts
