+++ react/lib/components/PayButton/PayButton.stories.tsx
