+++ react/lib/components/PayButton/index.ts
