+++ react/lib/components/PaymentDialog/PaymentDialog.tsx
