+++ react/lib/components/PaymentDialog/index.ts
