+++ react/lib/components/Widget/WidgetContainer.tsx
