+++ react/lib/components/Widget/index.ts
