+++ react/lib/typings.d.ts
