+++ react/lib/util/randomizeSats.ts
