+++ react/tsconfig.json
@@ -9,7 +9,7 @@
     "moduleResolution": "node",
     "jsx": "react",
     "sourceMap": true,
-    "declaration": true,
+    "declaration": false,
     "esModuleInterop": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
@@ -28,7 +28,8 @@
     "noEmit": true
   },
   "include": [
-    "src"
+    "src",
+    "lib",
   ],
   "exclude": [
     "node_modules",
